+++ CHANGELOG.md
@@ -13,13 +13,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     (:class:`.bdd.Utilisation`, :class:`.bdd.BaseCiblage` and
     :class:`.bdd.Ciblage`) and enums (:class:`.bdd.CibleType` and
     :class:`.bdd.UtilEtat`);
+  - Updated ``!fillroles`` to fill BaseCiblages too (new config option
+    :attr:`config.max_ciblages_per_action`);
   - New convenience function :func:`.bdd.base.autodoc_DynamicOneToMany`
     for documenting dynamicly loaded one-to-many relationships;
 
 ### Changed
 
   - Updated existing data classes attributes to link to new tables;
-  - New option ``nullable`` for :func:`.bdd.base.autodoc_DynamicOneToMany`.
+  - New option ``nullable`` for :func:`.bdd.base.autodoc_ManyToOne`;
+  - ``!fillroles`` now post camps descriptions in roles channel;
+  - Extended :attr:`.bdd.model_jeu.Camp.description` max length to 1000.
 
 ### Fixed